moved data spec to data file
