+++ src/aioway/spaces/README.md
@@ -1,3 +1,5 @@
 # `Space`s
 
-Spaces are essentially contracts to how a `torch.Tensor` should behave.
+Spaces are types that impose additional constraints than `isinstance`.
+
+
