+++ src/aioway/spaces/__init__.py
@@ -1,6 +1,7 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .casts import *
+from .data import *
 from .dims import *
 from .gym import *
 from .media import *
