+++ src/aioway/spaces/spaces.py
@@ -7,9 +7,8 @@ import dataclasses as dcls
 import typing
 
 from aioway._api import public_api
-from aioway._utils import torch_fake_mode
 
-__all__ = ["Space", "SpaceLike", "DataSpace", "AnySpace", "space_dcls"]
+__all__ = ["Space", "SpaceLike", "AnySpace", "space_dcls"]
 
 
 @public_api
@@ -55,19 +54,7 @@ class Space[T = typing.Any](abc.ABC):
 
 @public_api
 @space_dcls
-class DataSpace[T = typing.Any](Space[T], abc.ABC):
-    def sample(self, n: int = 1) -> T:
-        with torch_fake_mode():
-            return self._sample_n(n)
-
-    @abc.abstractmethod
-    def _sample_n(self, n: int, /) -> T:
-        raise NotImplementedError
-
-
-@public_api
-@space_dcls
-class AnySpace(DataSpace):
+class AnySpace(Space):
     """
     A `Space` that imposes no constraints.
     """
@@ -75,7 +62,3 @@ class AnySpace(DataSpace):
     @typing.override
     def contains(self, value):
         return True
-
-    @typing.override
-    def _sample_n(self, batch_size: int):
-        return object()
