+++ src/aioway/spaces/tdicts.py
@@ -14,7 +14,8 @@ import tensordict as td
 from aioway._api import public_api
 from aioway._utils import is_real_tensor, torch_fake_mode
 
-from .spaces import DataSpace, space_dcls
+from .data import DataSpace
+from .spaces import space_dcls
 from .tensors import Attr, AttrLike, DType
 
 __all__ = ["Schema", "TdictSpace", "TensorClassSpace"]
