--- src/aioway/specs/README.md
@@ -1,5 +1,0 @@
-# Specs
-
-Where a `Space` defines how a `torch.Tensor` / `td.TensorDict` should look like, a `Spec` defines how a `nn.Module` should behave.
-
-`Space` for data is akin to `Spec` for functions.
