--- src/aioway/specs/__init__.py
@@ -1,1 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
